+++ CHANGELOG.md
@@ -1,4 +1,14 @@
 
+## 1.1.0
+September 10, 2022
+- AppState class now with controller, materialApp, and cupertinoApp parameter
+- AppStatefulWidget class now with circularProgressIndicator parameter
+- Updated README.MD
+
+## 1.0.0
+August 24, 2022
+- Production release 
+
 ## 0.2.0
 August 21, 2022
 - import '../example/test/widget_test.dart';